申込可能な稼働期間がないとき、2020年内にページを開いた場合には2021年サービススタートの旨を表示するようにした
