+++ public/ke_ban.html
@@ -38,7 +38,7 @@
 
     <div id="unavailable" class="d-none">
       <div>
-        <p>
+        <p id="unavailable-message">
           現在申込可能な稼働期間はありません。
         </p>
         <p>
