+++ src/ke_ban_entry.js
@@ -1,3 +1,5 @@
+import "../public/rollbar/rollbar";
+
 import $ from "jquery";
 import dayjs from "dayjs";
 import "dayjs/locale/ja";
@@ -6,7 +8,7 @@ import {
     TODAY,
     get_terms_prev_now_next,
     get_available_terms
-} from "./ke_ban";
+} from "./logics/ke_ban";
 
 $(() => {
     const terms = get_terms_prev_now_next(TODAY);
@@ -23,6 +25,11 @@ $(() => {
     if (display_terms.length > 0) {
         document.getElementById("available_term").innerText = `${display_terms.join("\r\n")}`;
     } else {
-        document.getElementById("available_term").innerText = "現在お申込み頂ける稼働期間はございません。";
+        let message = "現在お申込み頂ける稼働期間はございません。";
+        // 2020年内の場合は1月スタートである旨を補足
+        if (TODAY.year() < 2021) {
+            message += "\r\n※2021年01月01日から本サービスをご利用頂けます";
+        }
+        document.getElementById("available_term").innerText = message;
     }
 });
